Splitting declaration builders across different files
