+++ Sources/TestCommons/DeclarationBuilder.swift
@@ -1,0 +1,4 @@
+/// Base class for builders for types, members and global variables and functions.
+public class DeclarationBuilder<T> {
+    
+}
