+++ Sources/TestCommons/IntentionCollectionBuilder.swift
@@ -303,38 +303,6 @@ public class FileIntentionBuilder {
     }
 }
 
-public protocol _FunctionBuilder {
-    associatedtype FunctionType = FunctionIntention
-    var target: FunctionType { get nonmutating set }
-}
-
-public extension _FunctionBuilder where FunctionType: MutableSignatureFunctionIntention {
-    var signature: FunctionSignature { get { return target.signature } nonmutating set { target.signature = newValue } }
-    
-    // TODO: It is less than ideal to repeat the name argument here, since most
-    // likely the user got here from function builders that already require a
-    // name during their creation.
-    @discardableResult
-    func createSignature(name: String, _ builder: (FunctionSignatureBuilder) -> Void) -> Self {
-        let b = FunctionSignatureBuilder(signature: FunctionSignature(name: name))
-        
-        builder(b)
-        
-        target.signature = b.build()
-        
-        return self
-    }
-}
-
-extension _FunctionBuilder where FunctionType: MutableFunctionIntention {
-    @discardableResult
-    public func setBody(_ body: CompoundStatement) -> Self {
-        target.functionBody = FunctionBodyIntention(body: body)
-        
-        return self
-    }
-}
-
 public class FunctionSignatureBuilder {
     var signature: FunctionSignature
     
@@ -390,370 +358,6 @@ public class FunctionBuilder<T: FunctionIntention>: _FunctionBuilder {
     }
 }
 
-public class MemberBuilder<T: MemberGenerationIntention> {
-    var targetMember: T
-    
-    public init(targetMember: T) {
-        self.targetMember = targetMember
-    }
-    
-    @discardableResult
-    public func addHistory(tag: String, description: String) -> MemberBuilder {
-        targetMember.history.recordChange(tag: tag, description: description)
-        
-        return self
-    }
-    
-    @discardableResult
-    public func setAccessLevel(_ accessLevel: AccessLevel) -> MemberBuilder {
-        targetMember.accessLevel = accessLevel
-        return self
-    }
-    
-    @discardableResult
-    public func addSemantics<S: Sequence>(_ semantics: S) -> MemberBuilder where S.Element == Semantic {
-        targetMember.semantics.formUnion(semantics)
-        return self
-    }
-    
-    @discardableResult
-    public func addAnnotations(_ annotations: [String]) -> MemberBuilder {
-        targetMember.annotations.append(contentsOf: annotations)
-        return self
-    }
-    
-    @discardableResult
-    public func addAttributes(_ attributes: [KnownAttribute]) -> MemberBuilder {
-        targetMember.knownAttributes.append(contentsOf: attributes)
-        return self
-    }
-    
-    public func build() -> T {
-        return targetMember
-    }
-}
-
-public extension MemberBuilder where T: OverridableMemberGenerationIntention {
-    @discardableResult
-    func setIsOverride(_ isOverride: Bool) -> MemberBuilder {
-        targetMember.isOverride = isOverride
-        return self
-    }
-}
-
-public extension MemberBuilder where T: InitGenerationIntention {
-    @discardableResult
-    func setIsConvenience(_ isConvenience: Bool) -> MemberBuilder {
-        targetMember.isConvenience = isConvenience
-        return self
-    }
-}
-
-extension MemberBuilder: _FunctionBuilder where T: FunctionIntention {
-    public typealias FunctionType = T
-    
-    public var target: FunctionType { get { return targetMember } set { targetMember = newValue }}
-}
-
-public extension MemberBuilder where T: MutableValueStorageIntention {
-    @discardableResult
-    func setValueStorage(_ storage: ValueStorage) -> MemberBuilder {
-        targetMember.storage = storage
-        
-        return self
-    }
-    
-    @discardableResult
-    func setOwnership(_ ownership: Ownership) -> MemberBuilder {
-        targetMember.storage.ownership = ownership
-        
-        return self
-    }
-    
-    @discardableResult
-    func setStorageType(_ type: SwiftType) -> MemberBuilder {
-        targetMember.storage.type = type
-        
-        return self
-    }
-    
-    @discardableResult
-    func setIsConstant(_ isConstant: Bool) -> MemberBuilder {
-        targetMember.storage.isConstant = isConstant
-        
-        return self
-    }
-    
-    @discardableResult
-    func setInitialValue(_ expression: Expression?) -> MemberBuilder {
-        targetMember.initialValue = expression
-        
-        return self
-    }
-}
-
-public extension MemberBuilder where T: PropertyGenerationIntention {
-    @discardableResult
-    func setAsField() -> MemberBuilder {
-        targetMember.mode = .asField
-        
-        return self
-    }
-    
-    @discardableResult
-    func setSetterAccessLevel(_ accessLevel: AccessLevel?) -> MemberBuilder {
-        targetMember.setterAccessLevel = accessLevel
-        
-        return self
-    }
-    
-    @discardableResult
-    func setAsComputedProperty(body: CompoundStatement) -> MemberBuilder {
-        targetMember.mode = .computed(FunctionBodyIntention(body: body))
-        
-        return self
-    }
-    
-    @discardableResult
-    func setAsGetterSetter(getter: CompoundStatement,
-                                  setter: PropertyGenerationIntention.Setter) -> MemberBuilder {
-        
-        targetMember.mode = .property(get: FunctionBodyIntention(body: getter),
-                                      set: setter)
-        
-        return self
-    }
-    
-    @discardableResult
-    func setInitialValue(expression: Expression?) -> MemberBuilder {
-        targetMember.mode = .computed(FunctionBodyIntention(body: []))
-        
-        return self
-    }
-    
-    @discardableResult
-    func setIsStatic(_ isStatic: Bool) -> MemberBuilder {
-        if isStatic {
-            if !targetMember.attributes.contains(.attribute("class")) {
-                targetMember.attributes.append(.attribute("class"))
-            }
-        } else {
-            targetMember.attributes.removeAll(where: { $0 == .attribute("class") })
-        }
-        
-        return self
-    }
-}
-
-public class TypeBuilder<T: TypeGenerationIntention> {
-    var targetType: T
-    
-    public init(targetType: T) {
-        self.targetType = targetType
-    }
-    
-    @discardableResult
-    public func addHistory(tag: String, description: String) -> TypeBuilder {
-        targetType.history.recordChange(tag: tag, description: description)
-        
-        return self
-    }
-    
-    @discardableResult
-    public func createProperty(named name: String,
-                               type: SwiftType,
-                               attributes: [PropertyAttribute] = []) -> TypeBuilder {
-        
-        return createProperty(named: name, type: type, mode: .asField, attributes: attributes)
-    }
-    
-    @discardableResult
-    public func createProperty(named name: String,
-                               type: SwiftType,
-                               mode: PropertyGenerationIntention.Mode = .asField,
-                               attributes: [PropertyAttribute] = [],
-                               builder: (MemberBuilder<PropertyGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        let storage = ValueStorage(type: type, ownership: .strong, isConstant: false)
-        
-        let prop: PropertyGenerationIntention
-        
-        if targetType is ProtocolGenerationIntention {
-            prop = ProtocolPropertyGenerationIntention(name: name,
-                                                       storage: storage,
-                                                       attributes: attributes)
-        } else {
-            prop = PropertyGenerationIntention(name: name,
-                                               storage: storage,
-                                               attributes: attributes)
-        }
-        
-        prop.mode = mode
-        
-        let mbuilder = MemberBuilder(targetMember: prop)
-        
-        builder(mbuilder)
-        
-        targetType.addProperty(mbuilder.build())
-        
-        return self
-    }
-    
-    @discardableResult
-    public func createConstructor(withParameters parameters: [ParameterSignature] = [],
-                                  builder: (MemberBuilder<InitGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        let ctor = InitGenerationIntention(parameters: parameters)
-        let mbuilder = MemberBuilder(targetMember: ctor)
-        
-        builder(mbuilder)
-        
-        targetType.addConstructor(mbuilder.build())
-        
-        return self
-    }
-    
-    @discardableResult
-    public func createVoidMethod(
-        named name: String, builder: (MemberBuilder<MethodGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        let signature = FunctionSignature(name: name, parameters: [])
-        
-        return createMethod(signature, builder: builder)
-    }
-    
-    @discardableResult
-    public func createMethod(named name: String,
-                             returnType: SwiftType = .void,
-                             parameters: [ParameterSignature] = [],
-                             isStatic: Bool = false,
-                             builder: (MemberBuilder<MethodGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        let signature = FunctionSignature(name: name,
-                                          parameters: parameters,
-                                          returnType: returnType,
-                                          isStatic: isStatic)
-        
-        return createMethod(signature, builder: builder)
-    }
-    
-    @discardableResult
-    public func createMethod(_ signatureString: String,
-                             isStatic: Bool = false,
-                             builder: (MemberBuilder<MethodGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        var signature = try! FunctionSignatureParser.parseSignature(from: signatureString)
-        signature.isStatic = isStatic
-        
-        return createMethod(signature, builder: builder)
-    }
-    
-    @discardableResult
-    public func createMethod(_ signature: FunctionSignature,
-                             builder: (MemberBuilder<MethodGenerationIntention>) -> Void = emptyInit) -> TypeBuilder {
-        
-        let method: MethodGenerationIntention
-        
-        if targetType is ProtocolGenerationIntention {
-            method = ProtocolMethodGenerationIntention(signature: signature)
-        } else {
-            method = MethodGenerationIntention(signature: signature)
-        }
-        
-        if !(targetType is ProtocolGenerationIntention) {
-            method.functionBody = FunctionBodyIntention(body: [])
-        }
-        
-        let mbuilder = MemberBuilder(targetMember: method)
-        builder(mbuilder)
-        
-        targetType.addMethod(mbuilder.build())
-        
-        return self
-    }
-    
-    public func build() -> T {
-        return targetType
-    }
-}
-
-public extension TypeBuilder where T: ClassExtensionGenerationIntention {
-    
-    /// Marks the target type for this type builder as an implementation for a
-    /// category extension interface declaration.
-    @discardableResult
-    func setAsCategoryImplementation(categoryName: String) -> TypeBuilder {
-        targetType.categoryName = categoryName
-        
-        return self
-    }
-    
-}
-
-public extension TypeBuilder where T: InstanceVariableContainerIntention {
-    @discardableResult
-    func createInstanceVariable(named name: String, type: SwiftType) -> TypeBuilder {
-        let storage = ValueStorage(type: type, ownership: .strong, isConstant: false)
-        
-        let ivar = InstanceVariableGenerationIntention(name: name, storage: storage)
-        targetType.addInstanceVariable(ivar)
-        
-        return self
-    }
-}
-
-public extension TypeBuilder where T: BaseClassIntention {
-    
-    /// Marks the target type for this type builder as coming from an interface
-    /// declaration.
-    @discardableResult
-    func setAsInterfaceSource() -> TypeBuilder {
-        targetType.isInterfaceSource = true
-        
-        return self
-    }
-    
-    /// Marks the target type for this type builder as coming from a category
-    /// extension interface declaration.
-    @discardableResult
-    func setAsCategoryInterfaceSource() -> TypeBuilder {
-        targetType.isInterfaceSource = true
-        
-        return self
-    }
-    
-    @discardableResult
-    func createConformance(protocolName: String) -> TypeBuilder {
-        let prot = ProtocolInheritanceIntention(protocolName: protocolName)
-        targetType.addProtocol(prot)
-        
-        return self
-    }
-    
-    @discardableResult
-    func createSynthesize(propertyName: String, variableName: String? = nil) -> TypeBuilder {
-        let intent =
-            PropertySynthesizationIntention(
-                propertyName: propertyName,
-                ivarName: variableName ?? propertyName,
-                isExplicit: false,
-                type: .synthesize)
-        
-        targetType.addSynthesization(intent)
-        
-        return self
-    }
-}
-
-public extension TypeBuilder where T: ClassGenerationIntention {
-    @discardableResult
-    func inherit(from className: String) -> TypeBuilder {
-        targetType.superclassName = className
-        
-        return self
-    }
-}
-
 public class EnumTypeBuilder {
     var targetEnum: EnumGenerationIntention
     
@@ -772,43 +376,5 @@ public class EnumTypeBuilder {
     
     public func build() -> EnumGenerationIntention {
         return targetEnum
-    }
-}
-
-// MARK: - Typealiases
-public typealias PropertyBuilder = MemberBuilder<PropertyGenerationIntention>
-public typealias InstanceVarBuilder = MemberBuilder<InstanceVariableGenerationIntention>
-public typealias MethodBuilder = MemberBuilder<MethodGenerationIntention>
-public typealias InitializerBuilder = MemberBuilder<InitGenerationIntention>
-
-public extension MemberBuilder where T == PropertyGenerationIntention {
-    convenience init(name: String, type: SwiftType) {
-        let prop = PropertyGenerationIntention(name: name, type: type, attributes: [])
-        
-        self.init(targetMember: prop)
-    }
-}
-
-public extension PropertyGenerationIntention {
-    convenience init(name: String, type: SwiftType, builder: (PropertyBuilder) -> Void) {
-        self.init(name: name, type: type, attributes: [])
-        
-        builder(PropertyBuilder(targetMember: self))
-    }
-}
-
-public extension MethodGenerationIntention {
-    convenience init(name: String, builder: (MethodBuilder) -> Void) {
-        self.init(signature: FunctionSignature(name: name))
-        
-        builder(MethodBuilder(targetMember: self))
-    }
-}
-
-public extension InitGenerationIntention {
-    convenience init(builder: (InitializerBuilder) -> Void) {
-        self.init(parameters: [])
-        
-        builder(InitializerBuilder(targetMember: self))
     }
 }
